L1 client: `ethers` -> `alloy` (#2456)
* `ethers` rpc -> `alloy`

* forge fmt

* remove lock from sqlite build

* Revert "remove lock from sqlite build"

This reverts commit 98532577e8c386d757af92e8fb5235041d778c49.

* update sqlite lockfile

* Merge remote-tracking branch 'origin/main' into rm/ethers-rpc-to-alloy

* typo & `RESOLVE_LIMIT` comment

* typo ignore -> contract bindings

* remove `rm` of existing bindings

* fetch stake table test

* fix reference tests

* rewrite switching logic

* properly use polling interval

* don't rebuild ethers bindings

* allow `ethers` bindings to be checked in CI

* `fmt`

* Patch foundry with resolve limit fix

The patch can be remove when https://github.com/alloy-rs/core/pull/864
lands in a foundry release.

Use solc from nixpkgs v0.8.28 since that's up to date now.

* Revert "allow `ethers` bindings to be checked in CI"

This reverts commit bda2822443cc5d78f54c67b7d91a61392a44a956.

Hope we can generate the bindings on the CI again now.

* Generate bindings with patched foundry

Left a TODO comment about the need to specify linked libraries.
Something I still need to look into more closely because the address
would change in different environments.

* CI: contracts timeout 30 -> 120 min

We now need to compile foundry if it hasn't been previously pushed to
our cachix cache. This takes a while.

* Revert solc 0.8.28 -> 0.8.23

A more recent version of geth would be required to run with 0.8.28

Add some checks to the deployer code to ensure we are actually linking
libraries when we attempt to do so.

* bump todo_by by 1 month

The date has passed again leading to warnings/errors during compilation.

* fix typo

* `switching` -> `shutting_down`

---------

Co-authored-by: Mathis Antony <sveitser@gmail.com>