WIP: simplify data flow for epoch queues
- Some renames to make it easier to read.
- Mutate the state of the queue in place relying on EVM reverts to
  restore the state of the contract when the transactions revert.

Would have to still update abstract contract and tests, so mainly
pushing this for discussion. We would also have a new function for
clients to determine available epochs, if that's actually required.

I think for clients it might be fine if we just return the max of the
registrationEpoch variable in the contract and currentEpoch() + 1. If
the registration epoch gets bumped by one by the time their transaction
is included that's just a race condition we can't avoid anyway. They can
set a reasonable `validUntilEpoch` to avoid having to wait too long.