Deploy fee contract with alloy (#2547)
* Deploy fee contract with alloy

Adds a few functions to deploy contracts with alloy and instantiate
alloy bindings.

Note: the fee contract ethers bindings are still used in Espresso
bridge, to migrate the bridge code we need to also create a version of
EthSignatureKey that uses alloy types.

* Fix alloy contract deployment

Apparently we need to add some fillers for alloy to build a usable
contract transactions. Otherwise fields like from or nonce are missing.

* Fix local demo

The permissioned stake table was not being deployed because its address
was read from the env var. Due to changing transaction counts this then
changed the light client contract address. I'm not quite sure why this
only caused issues now.

- Remove some mentions of the long gone hotshot contract.

* Fix and speedup test

We need to remember to set low polling intervals when we're talking to
anvil, otherwise it's set to 7 seconds or something like that. I think
alloy has some logic to handle this automatically.