Conditionally compile application versions
For tests in their current state we will probably need to enable 2
versions most of the time. I'm not sure about configuration through
the environment, but this leads me to believe we can pass the proper
configuration to `rustc`.

    RUSTFLAGS="--cfg feature=\"fee\"" cargo +nightly -Z unstable-options rustc --print cfg

Note that `--print` option of `cargo rustc` is only available on nightly.

Closes #2558