Update sqlx (#2617)
Updating sqlx to `0.8.3` appears to resolve. Excessive memory usage when compiling bundled sqlite. It was originally planned to add `sqlite-unbundled` as well to use native/system sqlite, but apparently excessive memory usage disappears w/ out that change. And it may be disruptive as our deployments currently rely on the bundled sqlite. As a followup, we can consider adding our own `sqlite-unbundled` feature to easily avoid compiling sqlite where we know there is one available in testing.

---------

Co-authored-by: tbro <tbro@users.noreply.github.com>