Adds `sqlite-unbundled` feature (#2619)
Adds `sqlite-unbundled` to turn on `sqlx/sqlite-unbundled`. With this
enabled, native/system sqlite will be used. Avoiding the need to build
sqlite itself. This will probably be preferred usage in most dev/ test
cycles, as most systems have a way of providing a modern sqlite3, and
compiling it is slow. Feature `embedded-db` should continue to work as
before (with sqlx bundled sqlite), so this change shouldn't disrupt
any current operations.

Co-authored-by: tbro <tbro@users.noreply.github.com>