Use mold linker in nix env on Linux (#2628)
* nix: use mold on linux

The build and test CI workflows already use mold. So this change only
affects development in the nix environments on Linux systems.

* CI: ensure nix dev env works on ARM based macOS

* chore: fix flake.nix on MacOS

* don't use pkgs in overlay

I guess when the overlay is applied there is a `pkgs` in scope, but not
sure. Probably more idiomatic to refer to attributes of `prev` which is
the package set before applying the overlay.

* CI: use `just build` instead of custom command

Also covers the sequencer-sqlite crate.