enable stake table epochs (#2510)
* add currentEpoch logic

* consolidate nextRegistrationEpoch and nextExitEpoch into one function and add tests checking epochs

* set the expected epoch to be a constant instead of a calculation

* WIP: simplify data flow for epoch queues

- Some renames to make it easier to read.
- Mutate the state of the queue in place relying on EVM reverts to
  restore the state of the contract when the transactions revert.

Would have to still update abstract contract and tests, so mainly
pushing this for discussion. We would also have a new function for
clients to determine available epochs, if that's actually required.

I think for clients it might be fine if we just return the max of the
registrationEpoch variable in the contract and currentEpoch() + 1. If
the registration epoch gets bumped by one by the time their transaction
is included that's just a race condition we can't avoid anyway. They can
set a reasonable `validUntilEpoch` to avoid having to wait too long.

* update tests, mark overiddden functions as virtual so they can be overidden again if need be, update parent contract

* remove public getters in favor of public members

- Fix solhint.
- Add a test for an epoch lenght greater than one.
- Remove unnecessary return values from functions that push to queue.

* revert if hotshot blocks per epoch

* do not create a check for the current epoch = max uint64 as this rever^C automatically

* remove unneccesary comments

* uncomment the exitEscrowPeriod for validator exit and update tests

* restrict function to pure

---------

Co-authored-by: Mathis Antony <sveitser@gmail.com>