Use latest ubuntu LTS version as base for images (#2724)
* Use latest ubuntu LTS version as base for images

We build on ubuntu-latest on the CI which will upgrade automatically so
I think it makes sense to do so as well for our docker images.

I'm seeing errors of this style

    /bin/orchestrator: /lib/x86_64-linux-gnu/libc.so.6: version `GLIBC_2.39' not found (required by /bin/orchestrator)

to reproduce

    docker pull ghcr.io/espressosystems/espresso-sequencer/orchestrator:main
    docker run ghcr.io/espressosystems/espresso-sequencer/orchestrator:main

* fix: local docker build script

If we run with `--all-features` we build CUDA support for icicle which
is not something we actually want.

* dev env: use openssl 3.0

* dev-node: use sqlite instead of postgres