revert change to use openssl 3.0 in dev env
This causes some issues when linking against libcurl which is needed by
some of our rust code. So we need to do something smarter.

This will break the script to build docker containers locally with nix
but shouldn't break anything else.

We could use openssl_3 in the nix dev env everywhere but this would lead
to a huge number of nix derivations missing the upstream binary cache
and therefore would be really annoying.

I think this will unblock the CI and we can figure out what to do from
here.