Remove foundry patch (#2736)
* Remove foundry patch

The patch in alloy is now included in foundry.

- Create a separate dev shell with the legacy version of foundry that
  we need to generate ethers bindings.
- Switch to foundry.nix overlay because the nixpkgs version is likely to
  lag a lot because foundry doesn't properly version their code.
- Enable the optimizer. This used to be the default but this was changed
  in foundry recently.

* CI: nuke contracts rust cache if nix files change

I think this sometimes leads to weird compilation errors.

* use glob

* revert change to use openssl 3.0 in dev env

This causes some issues when linking against libcurl which is needed by
some of our rust code. So we need to do something smarter.

This will break the script to build docker containers locally with nix
but shouldn't break anything else.

We could use openssl_3 in the nix dev env everywhere but this would lead
to a huge number of nix derivations missing the upstream binary cache
and therefore would be really annoying.

I think this will unblock the CI and we can figure out what to do from
here.

* foundry: allow internal reverts

This is described here:

    https://book.getfoundry.sh/cheatcodes/expect-revert#error